+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.7.0] - 2023-03-23
+
+### Added
+
+- Added the `prefix` property to the `customElement` plugin's options.
+
+### Removed
+
+- Removed `tag` argument from the `Element` decorator.
+- Removed the `prefix` property from the `extract` plugin's options.
+
+### Updated
+
+- Updated the `tag` calculator logic.
+- Updated the `visualStudioCode` and `webTypes` plugins.
+
 ## [0.6.9] - 2023-03-17
 
 ### Updated